+++ mrmustard/__init__.py
@@ -56,6 +56,7 @@ class Settings:
         self.PROGRESSBAR = True
         self._seed = np.random.randint(0, 2**31 - 1)
         self.rng = np.random.default_rng(self._seed)
+        self.DEFAULT_BS_METHOD = "vanilla"  # can be 'vanilla' or 'schwinger'
 
     @property
     def SEED(self):